refactor(*): create error mixin and add assign it on base objects
this change adds a method to create an error that can be used by plugins without need to access the
playerError directly
