+++ src/base/playback.js
@@ -1,6 +1,6 @@
 import { extend } from './utils'
 import UIObject from './ui_object'
-import PlayerError from '../components/error'
+import ErrorMixin from './error_mixin'
 
 /**
  * An abstraction to represent a generic playback, it's like an interface to be implemented by subclasses.
@@ -87,34 +87,6 @@ export default class Playback extends UIObject {
    * @method stop
    */
   stop() {}
-
-  /**
-   * creates playback error.
-   * @method createError
-   * @param {Object} error should be an object with code, description, level and raw error.
-   * @return {Object} Object with formatted error data including origin and scope
-   */
-  createError(error) {
-    !this.name && (this.name = 'playback')
-    const defaultError = {
-      description: '',
-      level: PlayerError.Levels.FATAL,
-      origin: this.name,
-      scope: 'playback',
-      raw: {},
-    }
-
-    const errorData = Object.assign(defaultError, error, {
-      code: `${this.name}:${error && error.code || 'unknown'}`
-    })
-
-    if (this.playerError)
-      this.playerError.error(errorData)
-    else
-      Log.warn(this.name, 'PlayerError is not defined. Error: ', err)
-
-    return errorData
-  }
 
   /**
    * seeks the playback to a given `time` in seconds
@@ -239,6 +211,8 @@ export default class Playback extends UIObject {
     this.$el.remove()
   }
 }
+
+Object.assign(Playback.prototype, ErrorMixin)
 
 Playback.extend = function(properties) {
   return extend(Playback, properties)
