+++ src/base/ui_core_plugin.js
@@ -1,7 +1,10 @@
 import { extend } from './utils'
 import UIObject from './ui_object'
+import ErrorMixin from './error_mixin'
 
 export default class UICorePlugin extends UIObject {
+  get playerError() { return this.core.playerError }
+
   constructor(core) {
     super(core.options)
     this.core = core
@@ -36,6 +39,8 @@ export default class UICorePlugin extends UIObject {
     return this
   }
 }
+
+Object.assign(UICorePlugin.prototype, ErrorMixin)
 
 UICorePlugin.extend = function(properties) {
   return extend(UICorePlugin, properties)
